+++ apps/discover/src/pages/App.tsx
@@ -7,7 +7,7 @@ import { Providers as DiscoverProviders } from 'providers';
 import { isTest, Container, getTenantInfo } from '@cognite/react-container';
 import { SidecarConfig } from '@cognite/sidecar';
 
-import { SIDECAR } from 'constants/app';
+import { DiscoverSidecarConfig, SIDECAR } from 'constants/app';
 import ApplicationRoutes from 'core/routes';
 import { configureStore } from 'core/store';
 import { GlobalStyles } from 'styles/globalStyles';
@@ -25,51 +25,58 @@ export const AppRoot: React.FC = () => {
       environment: { tenant: possibleTenant, appName: SIDECAR.applicationId },
     })
   );
-  const [sidecar, setSidecar] = useState(SIDECAR);
+  const [sidecar, setSidecar] = useState<DiscoverSidecarConfig | undefined>(
+    undefined
+  );
 
   // overwrite fakeIdp user for e2e tests
   useEffect(() => {
     if (
       !process.env.REACT_APP_E2E_USER &&
-      sidecar.fakeIdp &&
-      sidecar.fakeIdp.length
+      SIDECAR.fakeIdp &&
+      SIDECAR.fakeIdp.length
     ) {
       fetch(`/uuid`)
         .then((res) => res.json())
         .then((res) => {
-          setSidecar((prevState) => ({
-            ...prevState,
-            fakeIdp: prevState.fakeIdp
-              ? prevState.fakeIdp.map((fakeIdp) => {
+          setSidecar(() => ({
+            ...SIDECAR,
+            fakeIdp: SIDECAR.fakeIdp
+              ? SIDECAR.fakeIdp.map((fakeIdp) => {
                   const isAdmin = fakeIdp.name?.toLowerCase().includes('admin');
                   return {
                     ...fakeIdp,
                     userId: (isAdmin ? 'admin-' : '') + res,
                   };
                 })
-              : prevState.fakeIdp,
+              : SIDECAR.fakeIdp,
           }));
         })
         .catch(() => {
           console.warn('UUID endpoint not accessible');
+          setSidecar(SIDECAR);
         });
+    } else {
+      setSidecar(SIDECAR);
     }
   }, []);
 
   // ReduxProvider should be moved to the react-container
   return (
     <ReduxProvider store={store}>
-      <Container sidecar={sidecar as SidecarConfig}>
-        <>
-          <DevelopmentHelpers />
-          <GlobalStyles />
-          {/* eg: feature flags, query cache */}
-          <DiscoverProviders>
-            {/* eg: app main entry point */}
-            <ApplicationRoutes project={possibleTenant} />
-          </DiscoverProviders>
-        </>
-      </Container>
+      {sidecar && (
+        <Container sidecar={sidecar as SidecarConfig}>
+          <>
+            <DevelopmentHelpers />
+            <GlobalStyles />
+            {/* eg: feature flags, query cache */}
+            <DiscoverProviders>
+              {/* eg: app main entry point */}
+              <ApplicationRoutes project={possibleTenant} />
+            </DiscoverProviders>
+          </>
+        </Container>
+      )}
     </ReduxProvider>
   );
 };