+++ packages/sidecar/src/index.ts
@@ -1,4 +1,9 @@
 export * from './getDefaultSidecar';
 export * from './generateUrl';
-export type { CDFCluster, FakeIdp, SidecarConfig } from './types';
+export type {
+  CDFCluster,
+  FakeIdp,
+  SidecarConfig,
+  CDFClusterSet,
+} from './types';
 export * from './mocks';